move map_flatten and search_is_some to their own modules
